Lmscan backend / agent

+++ modules/lmscan-backend/src/main/scala/io/leisuremeta/chain/lmscan/backend/entity/Account.scala
@@ -2,7 +2,7 @@ package io.leisuremeta.chain.lmscan.backend.entity
 
 final case class Account(
     address: String,
-    balance: Double,
-    amount: Double,
+    balance: BigDecimal,
+    amount: BigDecimal,
     createdAt: Long,
 )
